Refactor DataModel.cs into MVC-like structure and centralize constants
Issue #6: Split DataModel.cs into smaller organized files
- Created Models/Coordinates/ folder with Coordinate base class, Latitude, Longitude
- Created Models/Configuration/ folder with Preferences, CalculationOptions
- Created Models/Magnetic/ folder with MagneticModelSet, MagneticModel,
  MagneticModelCollection, Coefficients
- Created Models/Results/ folder with MagneticCalculations, MagneticValue, GeoMagVector
- Created Enums/GeoMagEnums.cs with CoordinateSystem, Algorithm, MagneticFieldUnit,
  knownModels enums
- Reduced DataModel.cs to a minimal compatibility reference file
- Maintained same GeoMagSharp namespace for full backward compatibility

Issue #10: Centralize magic numbers and constants
- Created GeoConstants.cs with organized constant sections:
  - File format constants (RecordLen, MaxDeg, MaxModules)
  - Geographic limits (MinLatitude, MaxLatitude, etc.)
  - WGS84 ellipsoid parameters (A2WGS84, B2WGS84)
  - Unit conversions (feet, meters, kilometers, miles)
  - Magnetic field unit conversions
- Updated Calculator.cs to use Constants.A2WGS84 and Constants.B2WGS84
  instead of hardcoded values

All 43 unit tests pass.

Closes #6
Closes #10

via [Happy](https://happy.engineering)
